# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Pegmatite Vault API
  version: 0.1.0
  description: 鉱物標本コレクションを管理するローカルWebアプリケーションAPI
servers:
  - url: http://localhost:8000/api
    description: ローカル開発環境
...
paths:
...
  /specimens:
    get:
      tags: [Specimens]
      summary: 標本一覧を取得
      operationId: listSpecimens
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: q
          in: query
          description: 標本番号、標本名、鉱物名、採集地名の部分一致検索
          schema: { type: string }
        - name: mineral_id
          in: query
          schema: { type: string, format: uuid }
        - name: locality_id
          in: query
          schema: { type: string, format: uuid }
        - name: acquisition_method_id
          in: query
          schema: { type: integer, minimum: 1 }
        - name: favorite
          in: query
          schema: { type: boolean }
        - name: sort
          in: query
          schema:
            type: string
            enum: [created_at, specimen_no, specimen_name]
            default: created_at
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: 標本一覧
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecimenPage"
        "422":
          $ref: "#/components/responses/ValidationError"
    post:
      tags: [Specimens]
      summary: 標本を登録
      description: specimen_no未指定時は同じ採集地の最大値+1をAPIが割り当てる。
      operationId: createSpecimen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecimenCreate"
      responses:
        "201":
          description: 登録した標本
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecimenDetail"
        "422":
          $ref: "#/components/responses/ValidationError"
  /specimens/{specimen_id}:
    parameters:
      - $ref: "#/components/parameters/SpecimenId"
    get:
      tags: [Specimens]
      summary: 標本詳細を取得
      operationId: getSpecimen
      responses:
        "200":
          description: 標本詳細
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecimenDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Specimens]
      summary: 標本を更新
      operationId: updateSpecimen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecimenUpdate"
      responses:
        "200":
          description: 更新した標本
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecimenDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [Specimens]
      summary: 標本を削除
      description: 関連画像は削除せず論理アーカイブする。
      operationId: deleteSpecimen
      responses:
        "204":
          description: 削除成功
        "404":
          $ref: "#/components/responses/NotFound"
...
  /minerals/{mineral_id}/specimens:
    get:
      tags: [Minerals]
      summary: 鉱物に関連する標本を取得
      operationId: listMineralSpecimens
      parameters:
        - $ref: "#/components/parameters/MineralId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 関連標本
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecimenPage"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /localities/{locality_id}/specimens:
    get:
      tags: [Localities]
      summary: 採集地に関連する標本を取得
      operationId: listLocalitySpecimens
      parameters:
        - $ref: "#/components/parameters/LocalityId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 関連標本
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecimenPage"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 24 }
    SpecimenId:
      name: specimen_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    MineralId:
      name: mineral_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    LocalityId:
      name: locality_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    NotFound:
      description: 対象が存在しない
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: 入力値が不正
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, examples: [resource_not_found] }
            message: { type: string }
            details:
              type: [object, array, "null"]
    PageMeta:
      type: object
      required: [page, page_size, total]
      properties:
        page: { type: integer, minimum: 1 }
        page_size: { type: integer, minimum: 1 }
        total: { type: integer, minimum: 0 }
    SpecimenCreate:
      type: object
      required: [specimen_name]
      properties:
        specimen_no: { type: [integer, "null"], minimum: 1 }
        specimen_name: { type: string, minLength: 1, maxLength: 255 }
        locality_id: { type: [string, "null"], format: uuid }
        acquisition_method_id: { type: [integer, "null"], minimum: 1 }
        collection_date: { type: [string, "null"], format: date }
        features: { type: [string, "null"] }
        note: { type: [string, "null"] }
        favorite: { type: boolean, default: false }
        mineral_ids:
          type: array
          uniqueItems: true
          default: []
          items: { type: string, format: uuid }
    SpecimenUpdate:
      type: object
      minProperties: 1
      properties:
        specimen_no: { type: integer, minimum: 1 }
        specimen_name: { type: string, minLength: 1, maxLength: 255 }
        locality_id: { type: [string, "null"], format: uuid }
        acquisition_method_id: { type: [integer, "null"], minimum: 1 }
        collection_date: { type: [string, "null"], format: date }
        features: { type: [string, "null"] }
        note: { type: [string, "null"] }
        favorite: { type: boolean }
        mineral_ids:
          type: array
          uniqueItems: true
          items: { type: string, format: uuid }
    SpecimenSummary:
      type: object
      required: [id, specimen_no, specimen_name, favorite, mineral_names, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        specimen_no: { type: integer }
        specimen_name: { type: string }
        locality:
          oneOf:
            - $ref: "#/components/schemas/LocalitySummary"
            - type: "null"
        favorite: { type: boolean }
        mineral_names:
          type: array
          items: { type: string }
        thumbnail_image_id: { type: [string, "null"], format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    SpecimenDetail:
      allOf:
        - $ref: "#/components/schemas/SpecimenSummary"
        - type: object
          required: [minerals, images]
          properties:
            acquisition_method:
              oneOf:
                - $ref: "#/components/schemas/AcquisitionMethod"
                - type: "null"
            collection_date: { type: [string, "null"], format: date }
            features: { type: [string, "null"] }
            note: { type: [string, "null"] }
            minerals:
              type: array
              items:
                $ref: "#/components/schemas/MineralSummary"
            images:
              type: array
              items:
                $ref: "#/components/schemas/Image"
    SpecimenPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/SpecimenSummary"
    Image:
      type: object
      required:
        [id, original_filename, media_type, file_size, sort_order, status, created_at]
      properties:
        id: { type: string, format: uuid }
        specimen_id: { type: [string, "null"], format: uuid }
        archived_from_specimen_id: { type: [string, "null"], format: uuid }
        original_filename: { type: string }
        media_type:
          type: string
          enum: [image/jpeg, image/png, image/webp]
        file_size: { type: integer, minimum: 0 }
        caption: { type: [string, "null"] }
        sort_order: { type: integer, minimum: 0 }
        status: { type: string, enum: [active, archived] }
        archived_at: { type: [string, "null"], format: date-time }
        created_at: { type: string, format: date-time }
...
    MineralSummary:
      type: object
      required: [id, japanese_name]
      properties:
        id: { type: string, format: uuid }
        japanese_name: { type: string }
        english_name: { type: [string, "null"] }
        formula: { type: [string, "null"] }
...
    LocalitySummary:
      type: object
      required: [id, locality_name]
      properties:
        id: { type: string, format: uuid }
        locality_name: { type: string }
        alias_name: { type: [string, "null"] }
...
    AcquisitionMethod:
      type: object
      required: [id, name]
      properties:
        id: { type: integer, minimum: 1 }
        name: { type: string }
        description: { type: [string, "null"] }
...